todo_app/requirements.txt
asgiref==3.3.1
autopep8==1.5.5
cffi==1.14.5
cryptography==3.4.5
Django==3.1.6
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
mysqlclient==1.3.12
django-mysql==2.2.0
